+++ apps/heureka-next/CHANGELOG.md
@@ -1,5 +1,20 @@
 # heureka-next
 
+## 0.2.1
+
+### Patch Changes
+
+- 150b32d: adjust node and npm deps that the version is open to the top
+- Updated dependencies [150b32d]
+  - @cloudoperators/juno-messages-provider@0.2.2
+  - @cloudoperators/juno-ui-components@2.38.1
+
+## 0.2.0
+
+### Minor Changes
+
+- ad03fff: The app is now able to communicate with the GraphQL backend.
+
 ## 0.1.0
 
 ### Minor Changes